Added the new meme and mast analysis
added scripts for analysis of the RILseq results and creating Figure 4a, Table S4, Figure 6B and data for Figures 4b-d and S5 and Table S5. Also included an analysis of the gcvB sub experiment,
Change in the prepare_meme_from_summary script includes adding a overlapping functionality for different elements the are on the same target.
run_alot_of_meme_processes.py was only changes to use blatner names for all elements.